Adicionando autenticação nas Mutations e regras

+++ src/resolvers/Mutation.js
@@ -1,25 +1,68 @@
-import uuidv4 from 'uuid/v4'
+import bcrypt from 'bcryptjs'
+import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
     async createUser(parent, args, { prisma }, info) {
-        return await prisma.mutation.createUser({data: args.data}, info)
+        if(args.data.password.length < 8 ) {
+            throw new Error('Password must be 8 characters or longer.')
+        }
+        const password = await bcrypt.hash(args.data.password, 10)
+        const user = await prisma.mutation.createUser({
+            data: {
+                ...args.data,
+                password
+            }
+        })
+
+        return {
+            user,
+            token: jwt.sign({userId: user.id }, 'thisisasecret')
+        }
     },
-    async deleteUser(parent, args, { prisma }, info) {
+    async login(parent, args, { prisma }, info) {
+        const user = await prisma.query.user({
+            where: {
+                email: args.data.email
+            }
+        })
+
+        if(!user){
+            throw new Error('Unable to login!')
+        }
+        const isMatch = await bcrypt.compare(args.data.password, user.password)
+
+        if(!isMatch){
+            throw new Error('Unable to login')
+        }
+
+        return {
+            user,
+            token: jwt.sign({userId: user.id }, 'thisisasecret')
+        }
+    },
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return await prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)        
+
         return await prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
     },
-    async createPost(parent, args, { prisma }, info) {
+    async createPost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return await prisma.mutation.createPost({
             data: {
                 title: args.data.title,
@@ -27,20 +70,58 @@ const Mutation = {
                 published: args.data.published,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }    
         }, info)
     },
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if(!postExists){
+            throw new Error('Unable to delete post')
+        }
+
         return await prisma.mutation.deletePost({
             where: {
                 id: args.id
             }
         }, info)
     },
-    async updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, request}, info) {
+        const userId = getUserId(request)
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        const isPublished = await prisma.exists.Post({
+            id: args.id,
+            published: true,
+        })
+
+        if(!postExists){
+            throw new Error('Unable to update post')
+        }
+
+        if(isPublished && args.data.published === false){
+            //Delete all comments
+            await prisma.mutation.deleteManyComments({
+                where: {
+                    post: {
+                        id: args.id
+                    }
+                }
+            })
+        }
         return await prisma.mutation.updatePost({
             where: {
                 id: args.id
@@ -48,13 +129,22 @@ const Mutation = {
             data: args.data
         }, info)
     },
-    async createComment(parent, args, { prisma }, info) {
+    async createComment(parent, args, { prisma, request}, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.data.post,
+            published: true
+        })
+
+        if(!postExists){
+            throw new Error('Unable to find post')
+        }
         return await prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -65,14 +155,37 @@ const Mutation = {
             }
         }, info)
     },
-    async deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if(!commentExists){
+            throw new Error('Unable to delete comment')
+        }
+
         return await prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        if(!commentExists){
+            throw new Error('Unable to update comment')
+        }
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
